backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
django-cors-headers
django-filter
Pillow
